Extract method for validation publication into service
